Alteração do esqueleto do jogo para que ele funcione com as novas regras criadas
